% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMrcs.R
\name{GAMrcs}
\alias{GAMrcs}
\title{GAM model with restricted ubic splines function fitted on the log2 and IQR transformed prenatal concentration of Cu}
\usage{
GAMrcs(input_data, conf_list, sdq, expo_list, nknots)
}
\arguments{
\item{input_data}{A dataframe with one imputed dataset}

\item{conf_list}{A character vector specifying names of model adjustment factors (confounders)}

\item{sdq}{A string defining which SDQ score will be fitted}

\item{expo_list}{A character vector defining which exposures concentration will be fitted}

\item{nknots}{A scalar defining number of knots for cubic splines}
}
\value{
Return a list of GAM models for each exposure of interest
}
\description{
GAM model with restricted ubic splines function fitted on the log2 and IQR transformed prenatal concentration of Cu
}
