% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGAM.R
\name{PlotGAM}
\alias{PlotGAM}
\title{Plot GAMs for a list of exposure}
\usage{
PlotGAM(data, variable_list, residuals = FALSE)
}
\arguments{
\item{data}{A list of GAM model(s)}

\item{variable_list}{A data.frame containing a list of exposures for wich regressions will be run, first column must contain exposure names}

\item{residuals}{A logical indicating if to display residuals or not (default = FALSE)}
}
\value{
A panel of plots
}
\description{
Plot GAMs for a list of exposure
}
